generate_captcha/requirements.txt
# requirements.txt

Pillow>=9.5,<11
